+++ example_monitoring.py
@@ -128,6 +128,8 @@ worker_thread.start()
 thread = threading.Thread(target=report_websocket_status, args=(binance_websocket_api_manager,))
 thread.start()
 
+binance_websocket_api_manager.create_monitoring_socket_server()
+
 print("./binance_websocket_status.json will get refreshed every 20 seconds! so the tools/check_binance_websocket_api_"
       "manager check_command is able to pick it up and deliver it to the icinga daemon! the check_command returns "
       "UNKOWN if the file is older than 60 seconds...")