+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.3.2.dev (development stage/unreleased)
+### Added
+- get_monitoring_status_icinga: reconnects and update check
+- create_monitoring_socket_server
 
 ## 1.3.2
 ### Added